Review TODO statements, refactor json loader
